+++ src/nti/externalization/tests/test_singleton.py
@@ -6,12 +6,12 @@ from __future__ import division
 from __future__ import print_function
 
 from nti.externalization.singleton import SingletonDecorator
+from nti.externalization.singleton import Singleton
 from nti.externalization.tests import ExternalizationLayerTest
 
 from hamcrest import assert_that
 from hamcrest import is_
 from hamcrest import is_not
-from hamcrest import same_instance
 
 # disable: accessing protected members, too many methods
 # pylint: disable=W0212,R0904
@@ -26,15 +26,16 @@ class TestSingleton(ExternalizationLayerTest):
         X = SingletonDecorator('X', (object,), {})
 
         # No context
-        assert_that(X(), is_(same_instance(X())))
+        assert_that(X(), is_(X()))
 
         # context ignored
-        assert_that(X('context'), is_(same_instance(X('other_context'))))
+        assert_that(X('context'), is_(X('other_context')))
 
         # two contexts for the common multi-adapter case
         assert_that(X('context', 'request'),
-                    is_(same_instance(X('other_context', 'other_request'))))
+                    is_(X('other_context', 'other_request')))
 
+        # no instance variables
         x = X()
         with self.assertRaises(AttributeError):
             x.b = 1
@@ -45,11 +46,30 @@ class TestSingleton(ExternalizationLayerTest):
     def test_singleton_when_ancestor_is_singleton(self):
         X = SingletonDecorator('X', (object,), {})
         Y = SingletonDecorator('Y', (X,), {})
-        class Z(Y):
+
+        class Z(Y): # pylint:disable=no-init
             pass
 
-        assert_that(X(), is_(same_instance(X())))
-        assert_that(Y(), is_(same_instance(Y())))
-        assert_that(Z(), is_(same_instance(Z())))
-        assert_that(X(), is_not(same_instance(Z())))
-        assert_that(Y(), is_not(same_instance(Z())))
+        assert_that(X(), is_(X()))
+        assert_that(Y(), is_(Y()))
+        assert_that(Z(), is_(Z()))
+        assert_that(X(), is_not(Z()))
+        assert_that(Y(), is_not(Z()))
+
+    def test_custom_eq_hash_ne_ignored(self):
+
+        class X(Singleton):
+
+            def __eq__(self, other): # pragma: no cover
+                raise Exception
+
+            __ne__ = __eq__
+
+            def __hash__(self): # pragma: no cover
+                raise Exception
+
+
+
+        assert_that(X(), is_(X()))
+        assert_that(hash(X()), is_(hash(X())))
+        assert_that(X(), is_not(self))